pyproject.toml
[tool.poetry]
name = "dummymessenger"
version = "0.1.0"
description = ""
authors = ["Igornevertired <igornevertired@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
coloredlogs = "^15.0.1"
httpx = "^0.28.1"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
aiohttp = "^3.11.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
